refactor(acp): extract types/constants to src/acp-lifecycle/acp-lifecycle-types.ts (#4235 step 1)
